requirements.txt
Flask
pymongo
python-dotenv
python-3.10.12
